Rounds 16-39 alignment + fix tapscript OP_CODESEPARATOR opcode index tracking
Major Bitcoin Core alignment across all 11 crates (1905 tests, 0 failures):

- Fix critical tapscript bug: direct push opcodes (0x01-0x4b) were not counted
  in opcode_index due to `continue` skipping the increment. Bitcoin Core uses
  `for (; pc < pend; ++opcode_pos)` which increments unconditionally. This
  caused OP_CODESEPARATOR to record wrong positions, breaking Schnorr signature
  verification for covenant scripts (e.g. repeated OP_TUCK CHECKSIGVERIFY
  CODESEPARATOR patterns at block 925565).

- PSBT: TAP_TREE parsing with depth/leaf_ver validation, control block size
  check (33+32k), MuSig2 key types, BIP370 v2 support

- Consensus: negative output value check (CVE-2010-5139), snapshot magic 0xff,
  sigop cost 3-component (legacy×4 + P2SH×4 + witness×1), BIP54 accurate
  legacy sigop counting

- Wallet: fee estimation provider trait, CoinGrinder branch-and-bound,
  descriptor checksum/key origins/multipath, sendmany/listtransactions/
  gettransaction RPCs with correct Bitcoin Core parameter positions

- Script: 6 tapscript error codes, OP_CODESEPARATOR position fix

- RPCs: signrawtransactionwithkey DEFAULT sighash, lockunspent no-op fix,
  prioritisetransaction dummy validation, ping/disconnectnode, descriptor RPCs

+++ crates/rbtc-wallet/tests/bip341_tests.rs
@@ -199,7 +199,7 @@ fn bip341_key_path_sighash_vectors() {
             .unwrap()
             .iter()
             .map(|u| TxOut {
-                value: u["amountSats"].as_u64().unwrap(),
+                value: u["amountSats"].as_u64().unwrap() as i64,
                 script_pubkey: Script::from_bytes(decode_hex(u["scriptPubKey"].as_str().unwrap())),
             })
             .collect();
